[not for land yet] hacky fix for x.ndim followed by sub
Summary:

Currently there is a pattern which is not handled correctly by
FX graph mode quantization:

```
def forward(self, x):
    ndim = x.ndim
    # or add, mul, div, etc
    x = torch.sub(x, ndim)
    return x
```

The reason this does not work is as follows:
1. x.ndim becomes a getattr node
2. the real world type of x.ndim is an integer, but this is not known from the graph (yet)
3. binary ops such as `torch.sub` require quantization of inputs
4. the framework inserts an observer to observe the output of `ndim`
5. the observer fails because `ndim` is not a Tensor

For now, we hack a bandaid to unblock some teams, none of this is for
land.  We will have to think of a better fix which is landable (TBD).

Test Plan:

```
python test/test_quantization.py TestQuantizeFx.test_getattr_with_nontensor_result
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: c71e8ec7d95dad03c6cd784e23fc319f1fd13490
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53120