[not for land yet]: fix using size of quant layer in torch._assert
Summary:

Before this diff, if we had code lik

```
x = any_quant_layer(...)
x_size0 = x.size(0)
torch._assert(x_size_0 == 1)
```

The convert code would try to insert a dequantize after `x_size0`,
because it was a descendant of a quantized node and it was needed
for a non-quantized operation.  Since the actual type of the `size`
function output is an integer, this does not make sense.

For now, this is fixed as a one-off to unblock a customer.  In the
future, we may need to think more deeply about all the functions which
can return non-quantized types from quantized tensors and make sure
they are all covered.

Test Plan:

```
python test/test_quantization.py TestQuantizeFx.test_assert_on_size_after_quant_layer
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 5f94ee373bbf053bc8684cc7861ff104d68ce1f1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53187