[PyTorch] Change MaybeOwned tests to use intrusive_ptrs
Upcoming changes to `MaybeOwned<T>` will require that T is
pointer-sized and has alignment 2 or larger, with the intention of
limiting it to intrusive_ptr_target-ish types and optimizing based on
that.
This diff updates the tests to continue to build under these new
requirements; it is being sent separately to demonstrate that the
tests continue to work on the current implementation.

Differential Revision: [D27630289](https://our.internmc.facebook.com/intern/diff/D27630289/)

ghstack-source-id: 125981888
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55554