Modified batched residuals return of torch.linalg.lstsq
SciPy and NumPy operate only on non-batched input and return an
empty array with shape (0,) if rank(a) != n.
The behavior for non-batched inputs is NumPy and SciPy compatible and the
same result is computed.
For batched inputs, if any matrix in the batch has rank less than `n`,
then an empty tensor is returned.

ghstack-source-id: 2597e1505ac9134d73c540680127f3f50e82afe1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54722