ns for fx: move node I/O dtype mapping to be local instead of global
Summary:

To support shadows of custom functions, we need to allow user to
specify I/O type of the custom functions.

This PR is a cleanup in preparation for making the above happen.
We make the I/O dtype mappings be generated by a function instead
of a global variable. In the next PR, we will add a hook so user
can modify these mappings.

Test Plan:
```
python test/test_quantization.py TestFXNumericSuiteCoreAPIs
```

ghstack-source-id: 01fdc3482c900b2ada14e29034e7bda55222b076
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57021