ns for fx: allow user functions in shadowing
Summary:

Allows usage of user functions in NS shadow APIs. We expose the
i/o mapping to the user APIs, and thread them throughout the code.

Note: the format of the mapping is currently not the best.  Saving
improving that for a future PR.

Test Plan:
```
python test/test_quantization.py TestFXNumericSuiteCoreAPIs.test_user_defined_function
```

ghstack-source-id: da533119e39e2c3fd44f8681f7ea23610ac84ae9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57022