[BE][SyncBN] Avoid sync stats in eval mode
Pull Request resolved: https://github.com/pytorch/pytorch/pull/56982

SyncBatchNorm should behave as a regular BN layer in eval model, this
change ensures that this is the case.

In particular, the bug was when `track_running_stats=False`, `bn_training` would be set to True in eval mode, but this would trigger a collective sync in syncBN.

However, in eval mode syncBN should behave like a regular BN layer and not do this sync.

Closes https://github.com/pytorch/pytorch/issues/48988

Ensured with unittest that when used for inference on a single rank, stats sync is not triggered.
ghstack-source-id: 127544421

Differential Revision: [D27579297](https://our.internmc.facebook.com/intern/diff/D27579297/)